add select songs feature and handle some playlists operations

+++ app/src/main/java/com/k/sekiro/musico/playmusic/domain/model/PlaylistSong.kt
@@ -23,6 +23,6 @@ import androidx.room.ForeignKey
     ]
 )
 data class PlaylistSong(
-    val playlistId: Int,
-    val songId:Int
+    val playlistId: Long,
+    val songId: Long
 )
